+++ python/arcticdb/util/environment_setup.py
@@ -1,83 +1,55 @@
-"""
-Copyright 2025 Man Group Operations Limited
-
-Use of this software is governed by the Business Source License 1.1 included in the file licenses/BSL.txt.
-
-As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
-"""
-
-from abc import ABC, abstractmethod
 import copy
+from abc import ABC, abstractmethod
 from enum import Enum
 import logging
 import os
+import socket
 import tempfile
 import time
 import re
 import pandas as pd
 import numpy as np
-from typing import List, Union 
+from typing import Any, Dict, List, Union
 
 from arcticdb.arctic import Arctic
 from arcticdb.options import LibraryOptions
 from arcticdb.storage_fixtures.s3 import BaseS3StorageFixtureFactory, real_s3_from_environment_variables
-from arcticdb.util.utils import DFGenerator, ListGenerators, TimestampNumber 
+from arcticdb.util.utils import DFGenerator, ListGenerators, TimestampNumber
 from arcticdb.version_store.library import Library
 
-#ASV captures console output thus we create console handler
-logLevel = logging.INFO
-logger = logging.getLogger(__name__)
-logger.setLevel(logLevel)
-console_handler = logging.StreamHandler()
-console_handler.setLevel(logLevel)
-formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-console_handler.setFormatter(formatter)
-logger.addHandler(console_handler)
 
 ## Amazon s3 storage bucket dedicated for ASV performance tests
 AWS_S3_DEFAULT_BUCKET = 'arcticdb-asv-real-storage'
-
-## The tests will use shared/persistent storage with 3 different prefixes, creating 3 important STORAGE SPACES:
-## First - (PERSISTENT/PERMANENT STORAGE SPACE) This prefix is for production code - the libraries written there will 
-##    be SHARED across environments and product versions. Therefore there you place mainly READ ONLY libraries.
-##    Benefits:
-##       - created only once - no need to recreate them and waste time (this way we have time that is saved)
-##       - all read operations will be able ok to persist once their libraries
-##       - there is always a procedure that can recreate the store if needed
-##       - implicitly since those stores are generated ONCE we get coverage for risks of data not being 
-##         able to be read in future versions
-##       - as there should be always check data method we always assure things are there to recreate the storage
-##
-##    NOTE: What you should avoid is any modifications on his storage type
-##
-##    OPPORTUNITY: AGING TESTS - this is one class of tests that are now possible with shared store. The tests use
-##       the permanent store to append data each time they are executed, to add new symbols also. This gives 
-##       opportunity to have new writes, new appends in actual condition over time. Those types of tests will not 
-##       be flat line on graphs but will show how well or unwell we age on SPECIFIC environment - storage, because 
-##       the tests are always executed on same environment
 PERSISTENT_LIBS_PREFIX = "PERMANENT_LIBRARIES" 
-
-
-## Second - (MODIFIABLE STORAGE SPACE) When a test requires to write something to measure performance it needs
-##    a storage space different from the main permanent space. This is the modifyable space. There each test will
-##    need own library to run write and any other library modification tests.
-## 
-##    IMPORTANT: your test MIGHT run in several processes competing (or writing) for same library! As arcticdb has
-##          no locking or any other mechanism to prevent that you might end up with completely mixed results or
-##          or broken lib (which might go unnoticed also). Therefore you MUST:
-##            - create writable libraries for each process!
-##            - the test must reuse only the library of its own process!
 MODIFIABLE_LIBS_PREFIX = 'MODIFIABLE_LIBRARIES' 
-
-
-## Third - (MODIFIABLE STORAGE SPACE) How to test your test before going into production? How to experiment with 
-##    different values while debugging or researching safely without polluting or hurting the permanent libraries?
-##    this storage space resolves all such problems. A setup environment class in test mode will always write
-##    data to this storage space instead on permanent storage space. All that will be needed is to set 
-##    this class in test mode
 TEST_LIBS_PREFIX = 'TESTS_LIBRARIES' 
 
-class SetupConfig:
+def get_logger_for_asv(bencmhark_cls = None):
+    logLevel = logging.INFO
+    if bencmhark_cls:
+        logger = logging.getLogger(bencmhark_cls.__name__)
+    else:
+        logger = logging.getLogger("ASV")
+    logger.setLevel(logLevel)
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(logLevel)
+    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    console_handler.setFormatter(formatter)
+    logger.addHandler(console_handler)
+    return logger
+
+
+class Storage(Enum):
+    AMAZON = 1
+    LMDB = 2
+
+
+class LibraryType(Enum):
+    PERSISTENT = "PERMANENT"
+    MODIFIABLE = "MODIFIABLE"
+
+
+class StorageSetup:
     '''
     Defined special one time setup for real storages.
     Place here what is needed for proper initialization
@@ -89,7 +61,7 @@ class SetupConfig:
 
     def __new__(cls, *args, **kwargs):
         if not cls._instance:
-            cls._instance = super(SetupConfig, cls).__new__(cls, *args, **kwargs)
+            cls._instance = super(StorageSetup, cls).__new__(cls, *args, **kwargs)
             cls._instance.value = "Initialized"
 
             ## AWS S3 variable setup
@@ -100,11 +72,14 @@ class SetupConfig:
             cls._aws_default_factory.default_bucket = AWS_S3_DEFAULT_BUCKET
             cls._aws_default_factory.clean_bucket_on_fixture_exit = False
    
-    def get_aws_enforced_prefix():
+    def get_machine_id():
         """
-        Defined for test executors machines or individually
+        Returns machine id, or id specified through environments variable (for github)
         """
-        return os.getenv("ARCTICDB_PERSISTENT_STORAGE_SHARED_PATH_PREFIX", None)
+        id = os.getenv("ARCTICDB_PERSISTENT_STORAGE_SHARED_PATH_PREFIX", None)
+        if id is None:
+            id = socket.gethostname()
+        return  id
 
     @classmethod
     def get_aws_s3_arctic_uri(cls, prefix: str = MODIFIABLE_LIBS_PREFIX, confirm_persistent_storage_need: bool = False) -> str:
@@ -120,997 +95,150 @@ class SetupConfig:
         return cls._aws_default_factory.create_fixture().arctic_uri
 
 
-class Storage(Enum):
-    AMAZON = 1
-    LMDB = 2
+class StorageAccess:
 
-
-class LibraryType(Enum):
-    PERSISTENT = 1
-    MODIFIABLE = 2
-
-
-class StorageInfo:    
-
-    def __init__(self, storage: Storage, uris_cache: List[str], library_options: LibraryOptions, prefix: str) :
+    def __init__(self, storage: Storage, name_benchmark: str, library_options: LibraryOptions = None) :
+        """
+        Populate `name_benchamrk` to get separate modifiable space for 
+        """
         self.storage: Storage = storage
-        self.uris_cache: List[str] = uris_cache
         self.library_options: LibraryOptions = library_options
-        self.prefix: str = prefix
+        self.name_benchmark: str = name_benchmark
+        self._test_mode = False
+        self._uris_cache: List[str] = []
+        self._ac_cache = {}
+        StorageSetup()
 
-    def __str__(self):
-        value = super().__str__()
-        for key, val in vars(self).items():
-            value = f"{value}\n {key}: {val}"
-        return value
+    def log_info(self):
+        logger = get_logger_for_asv()
+        logger.info(f"{self} information:")
+        for key in self._ac_cache.keys():
+            logger.info(f"Arctic URI: {key}")
 
+    # Currently we're using the same arctic client for both persistant and modifiable libraries.
+    # We might decide that we want different arctic clients (e.g. different buckets) but probably not needed for now.
+    def _get_arctic_client(self) -> Arctic:
+        prefix = PERSISTENT_LIBS_PREFIX
+        if self._test_mode == True:
+            prefix = TEST_LIBS_PREFIX
+        return self.__get_arctic_client_internal(prefix, confirm_persistent_storage_need = True)
 
-class EnvConfigurationBase(ABC):
-    """
-    Defines base class for benchmark scenario setup and teardown
-    You need to create your own class inheriting from this one for each ASV benchmark class.
-    It will be responsible for setting up environment and providing tear down
+    def _get_arctic_client_modifiable(self) -> Arctic:
+        prefix = f"{MODIFIABLE_LIBS_PREFIX}_{StorageSetup.get_machine_id()}/{self.name_benchmark}"
+        return self.__get_arctic_client_internal(prefix, confirm_persistent_storage_need = False)
 
-    Override in child class:
-     (A) :func:`setup_all` - to setup required libraries and symbols (on empty storage that non exist/ do no deleted here)
-     (B) :func:`check_ok` - if the test will delete each time previous libraries and start from scratch 
-               then simply return False. Otherwise consider implementing logic that will check if needed
-               config exist on persistent store and if yes setup of configuration will be skipped
-     (C) OPTIONAL! :func:`setup_environment` this function should be used for `setup_cache` method of ASV test.
-               Should you need different behavior skip
-    
-    Use default :func:`setup_environment` for setup (primarily in setup_cache method of ASV class)
-
-    See :class:`GeneralSetupLibraryWithSymbols` for implementation of general logic 
-    See :class:`LMDBReadWrite` and :class:`AWSReadWrite` for test implementation and usage of environment configuration class
-
-    """
-
-    def __init__(self, storage: Storage, prefix: str = None, library_options: LibraryOptions = None, uris_cache: List[str] = None):
-        """
-        Should be initialized with type and arctic url
-        """
-        self.storage = storage
-        if uris_cache :
-            self.ac_cache = {keys: None for keys in uris_cache}
-        else:
-            self.ac_cache = {}
-        self.__storage_test_env_prefix: str = prefix
-        self.__libraries = set()
-        self.__library_options = library_options
-        self.__set_test_mode = False
-        SetupConfig()
-
-    def get_persistent_storage_prefix(self):
-        return PERSISTENT_LIBS_PREFIX # This should be same across environments - git, ec2, local etc
-
-    def get_modifiable_storage_prefix(self):
-        return self._combine_prefix(MODIFIABLE_LIBS_PREFIX)
-
-    def get_test_storage_prefix(self):
-        return self._combine_prefix(TEST_LIBS_PREFIX)
-
-    def _combine_prefix(self, type_prefix: str):
-        """
-        Constructs actual prefix based on enforced via 
-        environment variable + defined in the module type prefixes + defined by tests prefixes
-        all that makes unique combination
-        """
-        result = None
-        if SetupConfig.get_aws_enforced_prefix():
-            result = SetupConfig.get_aws_enforced_prefix()
-        if type_prefix:
-            if result is None:
-                result = type_prefix
-            else:
-                result = f"{result}_{type_prefix}"
-        if self.__storage_test_env_prefix:
-                result = f"{result}_{self.__storage_test_env_prefix}"
-        return result
-
-    def set_test_mode(self):
-        """
-        Makes the setup run in test mode. Will write data instead on the default persistent storage
-        in a separate test persistent storage space. This allows you to do test of production code 
-        and not pollute the default persistent store with unnecessary information and also protect data
-        there from deletion or corruption.
-        Always use test mode for new tests!
-        """
-        self.__set_test_mode=True
-        return self
-    
-    def is_test_mode(self) -> bool:
-        """
-        Checks if setup to work in test mode (in separate test storage space)
-        """
-        return TEST_LIBS_PREFIX in self.get_arctic_client_persistent().get_uri()
-
-    def logger(self) -> logging.Logger:
-        return logger
-        
-    @classmethod
-    def from_storage_info(cls, data: StorageInfo):
-        """
-        Constructs the class out of serialized data
-        """
-        return cls(storage=data.storage, uris_cache=data.uris_cache, library_options=data.library_options, prefix=data.prefix)
-    
-    def get_storage_info(self) -> StorageInfo:
-        """
-        Serializes the class information in order to be reconstructed back later
-        """
-        return StorageInfo(storage=self.storage, 
-                           uris_cache=list(self.ac_cache.keys()), library_options=self.__library_options, prefix=self.__storage_test_env_prefix)
-    
-    def _get_arctic_client(self, prefix, confirm_persistent_storage_need: bool = False):
-        """
-        Obtains specified client on specified persistent shared storage prefix,
-        allows creationg of test and modifiable etc clients
-        NOTE: prefix = `None` - the persistent storage where all persistent libraries will be created
-        """
+    def __get_arctic_client_internal(self, prefix, confirm_persistent_storage_need: bool = False) -> Arctic:
         if self.storage == Storage.AMAZON:
-            arctic_url = SetupConfig.get_aws_s3_arctic_uri(prefix, confirm_persistent_storage_need)
+            arctic_url = StorageSetup.get_aws_s3_arctic_uri(prefix, confirm_persistent_storage_need)
         elif self.storage == Storage.LMDB:
             ## We create fpr each object unique library in temp dir
             arctic_url = f"lmdb://{tempfile.gettempdir()}/benchmarks_{prefix}" 
-            ## 
-            # Home dir will be shared dir and will not work well with LMDB actually
-            #home_dir = os.path.expanduser("~")
-            #self.ac = Arctic(f"lmdb://{home_dir}/benchmarks")
         else:
             raise Exception("Unsupported storage type :", self.storage)
-        ac =  self.ac_cache.get(arctic_url, None)
+        ac =  self._ac_cache.get(arctic_url, None)
         if ac is None:
             ac = Arctic(arctic_url)    
-        self.ac_cache[arctic_url] = ac
-        return ac
-    
-    def get_arctic_client_persistent(self) -> Arctic:
-        if self.__set_test_mode:
-            return self._get_arctic_client(self.get_test_storage_prefix())
-        return self._get_arctic_client(self.get_persistent_storage_prefix(), confirm_persistent_storage_need=True)
-    
-    def get_arctic_client_modifiable(self) -> Arctic:
-        if self.__storage_test_env_prefix is None:
-            return self._get_arctic_client(self.get_modifiable_storage_prefix())
+        self._ac_cache[arctic_url] = ac
+        return ac    
+
+    def get_library_name(self, library_type, lib_name_suffix):
+        if library_type == LibraryType.PERSISTENT:
+            return f"{library_type}_{self.name_benchmark}_{lib_name_suffix}"
+        if library_type == LibraryType.MODIFIABLE:
+            # We want the modifiable libraries to be unique per process/ benchmark class. We embed this deep in the name
+            return f"{library_type}_{self.name_benchmark}_{os.getpid()}_{lib_name_suffix}"
+
+    def get_library(self, library_type : LibraryType, lib_name_suffix : str) -> Library:
+        lib_name = self.get_library_name(library_type, lib_name_suffix)
+        if library_type == LibraryType.PERSISTENT:
+           return self._get_arctic_client().get_library(lib_name, create_if_missing=True)
+        elif library_type == LibraryType.MODIFIABLE:
+           return self._get_arctic_client_modifiable().get_library(lib_name, create_if_missing=True, 
+                                                        library_options= self.library_options)
         else:
-            return self._get_arctic_client(self.get_modifiable_storage_prefix())
-    
-    def remove_all_modifiable_libraries(self, confirm=False):
-        """
-        Removes all libraries created on modifiable storage space
-        """
-        assert confirm, "Deletion of all libraries must be confirmed"
-        ac = self.get_arctic_client_modifiable()
-        libs = self.get_arctic_client_modifiable().list_libraries()
-        for lib in libs:
-            ac.delete_library(lib)
-    
-    def get_symbol_name_template(self, sym_suffix: Union[str, int]) -> List[str]:
-        """
-        Defines how the symbol name would be constructed
-        """
-        return f"sym_{sym_suffix}"
+            raise Exception(f"Unsupported library type: {library_type}")
 
-    def _get_library_names(self, lib_suffix: Union[str, int] = None) -> List[str]:
-        """
-        Redefine the way the names of libraries are constructed.
-        `lib_suffix` can be None indicating no lib suffix added,
-        or number - for instance number of symbols in library,
-        or a composite string
-        """
-        if self.__storage_test_env_prefix is None:
-            lib_mid_part_name = self.__class__.__name__
-        else: 
-            lib_mid_part_name = self.__storage_test_env_prefix
-        lib_names = [f"PERM_{lib_mid_part_name}_{lib_suffix}", 
-                     f"MOD_{lib_mid_part_name}_{lib_suffix}"]        
-        return lib_names
-    
-    def _get_lib(self, ac: Arctic, lib_name: str) -> Library:
-        """
-        Returns a library requested or constructs one if does not exist
-        keeps track of all libs requested
-        """
-        lib_opts = self.__library_options
-        self.__libraries.add(lib_name)
-        return ac.get_library(lib_name, create_if_missing=True, 
-                                library_options=lib_opts)
+    def has_library(self, library_type : LibraryType, lib_name_suffix : str) -> Library:
+        lib_name = self.get_library_name(library_type, lib_name_suffix)
+        if library_type == LibraryType.PERSISTENT:
+           return self._get_arctic_client().has_library(lib_name)
+        elif library_type == LibraryType.MODIFIABLE:
+           return self._get_arctic_client_modifiable().has_library(lib_name)        
+        else:
+            raise Exception(f"Unsupported library type: {library_type}")
 
-    def get_library_name(self, lib_type: LibraryType, library_suffix: Union[str, int] = None) -> str:
-        """
-        Returns library name on persistent space
-        """
-        if lib_type == LibraryType.PERSISTENT:
-            return self._get_library_names(library_suffix)[0]
-        else: 
-            return self._get_library_names(library_suffix)[1]
+    # Not sure how useful is this, one can always just keep track of the libraries created and clear them
+    def clear_all_modifiable_libs_from_this_process(self):
+        ac = self._get_arctic_client_modifiable()
+        lib_names = set(ac.list_libraries())
+        to_deletes = [lib_name for lib_name in lib_names 
+                      if f"_{os.getpid()}_" in lib_name]
+        for to_delete in to_deletes:
+            ac.delete_library(to_delete)
 
-    def get_library(self, library_suffix: Union[str, int] = None) -> Library:
-        """
-        Returns a library on persistent space. (created if do not exist)
-        The name is constructed internally, but user can add a suffix (index)
-        Uses `get_library_name` to construct the name
 
-        IMPORTANT: `get_library` will always return Library even if none exist will be created new!
-        """
-        return self._get_lib(self.get_arctic_client_persistent(), 
-                             self.get_library_name(LibraryType.PERSISTENT, library_suffix))
-
-    def has_library(self, library_suffix: Union[str, int] = None) -> bool:
-        """
-        checks if there is a library with that suffix on persistent storage space
-        """
-        return self.get_arctic_client_persistent().has_library(self.get_library_name(LibraryType.PERSISTENT, 
-                                                                                     library_suffix))
-
-    def get_modifiable_library(self, library_suffix: Union[str, int] = None) -> Library:
-        """
-        Returns library to read write and delete after done. (create if not exist is on!)
-        The name is constructed internally, but user can add a suffix (index)
-        Uses `get_library_name` to construct the name
-
-        IMPORTANT: `get_modifiable_library` will always return Library even if none exist will be created new!
-        """
-        return self._get_lib(self.get_arctic_client_modifiable(), 
-                             self.get_library_name(LibraryType.MODIFIABLE, library_suffix))
-
-    def has_modifiable_library(self, library_suffix: Union[str, int] = None) -> bool:
-        """
-        checks if there is a library with that suffix on persistent storage space
-        """
-        return self.get_arctic_client_persistent().has_library(self.get_library_name(LibraryType.MODIFIABLE, 
-                                                                                     library_suffix))
-    
-    def delete_modifiable_library(self, library_suffix: Union[str, int] = None):
-        """
-        Use this method to delete previously created library on modifiable storage 
-        space
-        NOTE: will assert if library is not delete
-        """
-        ac = self.get_arctic_client_modifiable()
-        name = self.get_library_name(LibraryType.MODIFIABLE, library_suffix)
-        logger.info(f"Deleting modifiable library {name}")
-        ac.delete_library(name)
-        assert name not in ac.list_libraries(), f"Library successfully deleted {name}"
-
-    def set_params(self, params):
-        """
-        Sets parameters for environment generator class. 
-        The expectations for parameters is same as ASV class.
-        To be used in concrete classes
-        """
-        self._params = params
-        return self
-
-    def get_parameter_list(self):
-        """
-        Returns the parameter list. Use it to set the `params` variable of ASV test
-        """
-        return self._params
-
-    def is_multi_params_list(self):
-        """
-        Checks if the parameters are list of 2+ parameter lists
-        """
-        return all(isinstance(i, list) for i in self._params)
-
+# It is quite clear what is this responsible for: only dataframe generation
+# Using such an abstraction can help us deduplicate the dataframe generation code between the different `EnvironmentSetup`s
+# Note: We use a class instead of a generator function to allow caching of dataframes in the state
+class DataFrameGenerator(ABC):
     @abstractmethod
-    def check_ok() -> bool:
-        """
-        Define mechanism to check if permanent setup is available
-        to skip the inital setup
-        """
-        pass
-    
-    @abstractmethod
-    def setup_all(self) -> 'EnvConfigurationBase':
-        '''
-        Provide implementation that will setup needed data assuming storage does not contain 
-        any previous data for this scenario
-        '''
+    def get_dataframe(self, **kwargs):
         pass
 
-    def setup_environment(self, library_sufixes_list: List[Union[str | int]] = None) -> 'EnvConfigurationBase':
+# The population policy is tightly paired with the `populate_library`
+# Using a separate abstraction for the population policy can allow us to be flexible with how we populate libraries.
+# E.g. One benchmark can use one population policy for read operations, a different for updates and a third for appends
+# Note: Using a kwargs generator to be passed to the df_generator allows customization of dataframe generation params (e.g. num_rows, num_cols, use_string_columns?)
+class LibraryPopulationPolicy(ABC):
+
+    def __init__(self, num_symbols: int, df_generator: DataFrameGenerator):
+        self.num_symbols = num_symbols
+        self.df_generator = df_generator
+
+    @abstractmethod
+    def get_symbol_name(self, ind: int):
+        pass
+
+    @abstractmethod
+    def get_generator_kwargs(self, ind: int) -> Dict[str, Any]:
+        pass
+
+    def populate_library(self, lib: Library, population_policy: 'LibraryPopulationPolicy'):
+        num_symbols = population_policy.num_symbols
+        df_generator = population_policy.df_generator
+        for i in range(num_symbols):
+            sym = population_policy.get_symbol_name(i)
+            kwargs = population_policy.get_generator_kwargs(i)
+            df = df_generator.get_dataframe(**kwargs)
+            lib.write(sym, df)
+
+    # This is the only API to populate a persistent library. If we deem useful we can also add a check whether library is valid (e.g. has the correct num_symbols)
+    # As discussed, ideally this will be called in completely separate logic from ASV tests to avoid races, but for simplicity
+    # for now we can just rely on setup_cache to populate the persistant libraries if they are missing.
+    def populate_persistent_library_if_missing(ac, benchmark_cls, lib_name_suffix, population_policy : 'LibraryPopulationPolicy'):
         """
-        Responsible for setting up environment.
-        Will first delete any pre-existing modifiable libraries
-        Then will check if persistent environment is there
-        If absent will create it, deleting firs any libraries that might exist initially        
-        if test is creating multiple libraries, provide 
-        `library_sufixes_list` to make sure all previous libs are cleaned
-
-        Use at :func:`setup_cache` method of ASV test classes
-        """
-        indexes = [None] # by default there will be one lib
-        if library_sufixes_list:
-            indexes = library_sufixes_list
-        for i in indexes:
-            self.delete_modifiable_library(i)
-        if not self.check_ok():
-            ac = self.get_arctic_client_persistent()
-            ## Delete all PERSISTENT libraries before setting up them
-            for i in indexes:
-                lib = self.get_library_name(LibraryType.PERSISTENT, i)
-                logger.info(f"REMOVING LIBRARY: {lib}")
-                ac.delete_library(lib)
-            self.setup_all()
-        return self
-    
-    @classmethod
-    def get_parameter_from_string(cls, string: str, param_indx: int, param_type: Union[int, str] = int) -> Union[int, str]:
-        """
-        A way to handle composite parameter list.
-
-        Instead of having multiple sets of parameter we keep the parameter to just one list
-        of string with encoded several values for 2 or more parameters delimited with "__"
-
-        For example following value contains 3 parameters: "w180__h200__value1234"
-
-        We can treat first parameter either as string or as integer with description:
-        - get_parameter_from_string('w180__h200__value1234', 0, str) will return 'w180' string
-        - get_parameter_from_string('w180__h200__value1234', 0, int) will extract int and return '180' as int value
-
-        """
-        params = string.split("__")
-        value = params[param_indx]
-        if param_type is str:
-            pass
-        elif param_type is int:
-            match = re.search(r'\d+', value)
-            if match:
-                value = int(match.group())
-        else:
-            raise Exception(f"Unsupported type {param_type}")
-        return value
-
-
-class SetupSingleLibrary(EnvConfigurationBase):
-    """
-    This class provides abstract logic for generation of a set of symbols
-    that have certain number of rows and optionally columns.
-
-    You have to use :func:`set_params` to define the exact profile of the symbols
-    to be created. The expectations are same as with ASV `params`: 
-         - a) a list of 2 lists = desired rows sizes (first) and column sizes (second)
-         - b) a list of 1 list - desired rows sizes (default column size to be used)
-
-    While in default implementation this class will generate list of random column types,
-    you can override that behavior in parent classes for specific needs    
-
-    """
-
-    def set_params(self, params) -> 'SetupSingleLibrary':
-        super().set_params(params)
-        return self
-
-    def generate_dataframe(self, rows: int, cols: int) -> pd.DataFrame:
-        """
-        Dataframe generator that will be used for wide dataframe generation
-        """
-        if cols is None:
-            cols = 1
-        st = time.time()
-        self.logger().info("Dataframe generation started.")
-        df: pd.DataFrame = DFGenerator.generate_random_dataframe(rows=rows, cols=cols)
-        self.logger().info(f"Dataframe rows {rows} cols {cols} generated for {time.time() - st} sec")
-        return df
-    
-    def get_symbol_name(self, rows, cols) -> str:
-        '''
-        Maps rows and cols to proper symbol name
-        This method should be used in ASV tests obtain symbol name based on parameters
-        '''
-        sym_suffix = ""
-        if cols is None:
-            sym_suffix = f"rows{rows}"
-        else:
-            sym_suffix = f"rows{rows}_cols{cols}"
-        return self.get_symbol_name_template(sym_suffix)
-    
-    def _get_symbol_bounds(self):
-        """
-        Extracts rows and cols from the parameter structure
+        lib_name = get_library_name(LibraryType.PERSISTENT, benchmark_cls, lib_name_suffix)
+        if ac.has_library(lib_name):
+            lib = ac.create_library(lib_name)
+            populate_library(lib, population_policy)            
         """
 
-        assert len(self._params) > 0, "No parameters defined"
 
-        if (self.is_multi_params_list()):
-            assert len(self._params) <= 2, "Supports 2 parameters maximum - rows and cols lists"
-            list_rows = self._params[0]
-            if len(self._params) == 2:
-                list_cols = self._params[1]
-            return (list_rows, list_cols)
-        else:
-            return (self._params, [ None ])
+sa = StorageAccess(Storage.AMAZON, "MY_NAME")
+print(sa._get_arctic_client().list_libraries())
+sa._test_mode = True
+sa.get_library(LibraryType.PERSISTENT, "haho")
+print(sa._get_arctic_client().list_libraries())
+sa.get_library(LibraryType.MODIFIABLE, "bebo")
+print(sa._get_arctic_client_modifiable().list_libraries())
+sa.log_info()
+sa.clear_all_modifiable_libs_from_this_process()
+print(sa._get_arctic_client_modifiable().list_libraries())
 
-    def setup_all(self) -> 'SetupSingleLibrary':
-        """
-        Sets up in default library a number of symbols which is combination of parameters
-        for number of rows and columns
-        For each of symbols it will generate dataframe based on `generate_dataframe` function
-        """
-        (list_rows, list_cols) = self._get_symbol_bounds()
-        start = time.time()
-        lib = self.get_library()
-        self.logger().info(f"Library: {lib}")
-        for row in list_rows:
-            for col in list_cols:
-                st = time.time()
-                df = self.generate_dataframe(row, col)
-                symbol = self.get_symbol_name(row, col)
-                lib.write(symbol, df)
-                self.logger().info(f"Dataframe stored at {symbol} for {time.time() - st} sec")
-        self.logger().info(f"TOTAL TIME (setup of one library with specified rows and column symbols): {time.time() - start} sec")
-        return self
-    
-    def check_ok(self):
-        """
-        Checks if library contains all needed data to run tests.
-        if OK, setting things up can be skipped
-        """
-        (list_rows, list_cols) = self._get_symbol_bounds()
-        if not self.has_library():
-            return False
-        lib = self.get_library()
-        symbols = set(lib.list_symbols())
-        self.logger().info(f"Symbols {symbols}")
-        for rows in list_rows:
-            for cols in list_cols:
-                symbol = self.get_symbol_name(rows, cols)
-                self.logger().info(f"Check symbol {symbol}")
-                if not symbol in symbols:
-                    return False
-        return True
 
 
-class SetupLibrariesWithVersionAndSnapshots(EnvConfigurationBase):
-    """
-    Will create several libraries each containing specified number of symbols.
-    Each symbol will have by default 1 version (use `set_max_number_versions` and
-    `set_mean_number_versions_per_sym` to override that) and will have no metadata
-    (use `set_with_metadata_for_each_version` to override that).
 
-    By default no snapshot will be generated for each version. This can be overridden with
-    `set_with_snapshot_for_each_version`
-    """
 
-    def __init__(self, storage, prefix = None, library_options = None, uris_cache = None):
-        super().__init__(storage, prefix, library_options, uris_cache)
-        self.with_metadata = False
-        self.versions_max = 1
-        self.mean = 1
-        self.with_snapshot = False
-        self.last_snapshot = None # Will hold the name of last snapshot created
-        self.first_snapshot = None # Will hold the name of first snapshot created
-        self.first_snapshot_taken = False
 
-    def set_params(self, params) -> 'SetupLibrariesWithVersionAndSnapshots':
-        super().set_params(params=params)
-        return self
 
-    def set_mean_number_versions_per_sym(self, mean) -> 'SetupLibrariesWithVersionAndSnapshots':
-        self.mean = mean
-        return self
 
-    def set_max_number_versions(self, versions_max) -> 'SetupLibrariesWithVersionAndSnapshots':
-        self.versions_max = versions_max
-        return self
-
-    def set_with_metadata_for_each_version(self, with_medatada: bool = True) -> 'SetupLibrariesWithVersionAndSnapshots':
-        self.with_metadata = with_medatada
-        return self
-    
-    def set_with_snapshot_for_each_version(self, with_snapshot: bool = True) -> 'SetupLibrariesWithVersionAndSnapshots':
-        self.with_snapshot = with_snapshot
-        return self
-
-    def generate_dataframe(self, num_rows):
-        return DFGenerator.generate_random_dataframe(rows=num_rows, cols=20)
-    
-    def generate_metadata(self):
-        return DFGenerator.generate_random_dataframe(rows=3, cols=10).to_dict()
-
-    def get_symbol_name(self, number_symbols: int, number_versions:int, with_metadata:bool, with_snapshot: bool) -> str:
-        '''
-        Maps rows and cols to proper symbol name
-        This method should be used in ASV tests obtain symbol name based on parameters
-        '''
-        if with_metadata:
-            meta = "with-meta"
-        else:
-            meta = "no-meta"
-        if with_snapshot:
-            snap = "with-snap"
-        else:
-            snap = "no-snap"
-        return self.get_symbol_name_template(f"{number_symbols}_vers-{number_versions}_{meta}_{snap}")
-    
-    def get_versions_list(self, number_symbols: int) -> List[np.int64]:
-        if self.versions_max == 1:
-            versions_list = [1] * number_symbols
-        else:
-            versions_list = ListGenerators.generate_random_list_with_mean(
-                number_elements=number_symbols,
-                specified_mean=self.mean, 
-                value_range=(1, self.versions_max), 
-                seed=365)
-        return versions_list
-
-    def setup_library(self, number_symbols) -> 'SetupLibrariesWithVersionAndSnapshots':
-        """
-        Sets a library with specified number of symbols with a snapshot and metadata if specified
-        """
-        lib = self.get_library(number_symbols)
-        df = self.generate_dataframe(10)
-        meta = self.generate_metadata()
-        self.logger().info(f"METADATA: {meta}")
-        versions_list = self.get_versions_list(number_symbols)
-        index = 0
-        for sym_num in range(number_symbols):
-            versions = versions_list[index]
-            symbol = self.get_symbol_name(number_symbols=sym_num,
-                                       number_versions=versions,
-                                       with_metadata=self.with_metadata,
-                                       with_snapshot=self.with_snapshot)
-            self.logger().info(f"Generating symbol {symbol}.")
-            st = time.time()
-            for num in range(versions):
-                if self.with_metadata:
-                    lib.write(symbol, df, metadata=meta)
-                else:
-                    lib.write(symbol, df)
-            if self.with_snapshot:
-                snap = f"snap_{symbol}"
-                if self.with_metadata:
-                    lib.snapshot(snap, metadata=meta)
-                    self.last_snapshot = snap # Last snapshot created
-                    if not self.first_snapshot_taken:
-                        self.first_snapshot = snap # Remember first snapshot
-                        self.first_snapshot_taken = True
-                else: 
-                    lib.snapshot(snap)
-            self.logger().info(f"Generation of {symbol} symbol COMPLETED for :{time.time() - st} sec")
-            index += 1
-        return self
-
-    def setup_all(self):
-        assert not self.is_multi_params_list(), "One parameters list expected"
-        st = time.time()
-        for num_symbols in self.get_parameter_list():
-            self.setup_library(num_symbols)
-        self.logger().info(f"Total time {time.time() - st}")    
-
-    def check_ok(self) -> bool:
-
-        for num_symbols in self.get_parameter_list():
-            if not self.has_library(num_symbols):
-                return False
-            lib = self.get_library(num_symbols)
-            symbols = set(lib.list_symbols())
-            self.logger().info(f"Check library: {lib}")
-            versions_list = self.get_versions_list(num_symbols)
-            index = 0
-            for sym_num in range(num_symbols):
-                symbol = self.get_symbol_name(number_symbols=sym_num,
-                            number_versions=versions_list[index],
-                            with_metadata=self.with_metadata,
-                            with_snapshot=self.with_snapshot)
-                if not (symbol in symbols):
-                    self.logger().info(f"Symbol {symbol} not found")
-                    return False
-                index += 1
-
-        return True        
-    
-    def clear_symbols_cache(self):
-        for num_symbols in self.get_parameter_list():
-            lib = self.get_library(num_symbols)
-            lib._nvs.version_store._clear_symbol_list_keys()
-
-
-class NoSetup(EnvConfigurationBase):
-    """
-    Provides only access to persistent storage space as well
-    as functionalities around work with modifiable storage space - 
-    create library and delete library `check_ok` always returns true, 
-    and `setup_all` does nothing but returns `StorageInfo`
-
-    Use this when you need to do custom work
-    """
-
-    def check_ok(self) -> bool:
-        return True
-
-    def setup_all(self) -> StorageInfo:
-        return self.get_storage_info()
-    
-
-class AppendDataSetupUtils(NoSetup):
-    """
-    This class presents general purpose utilities to generate sequenced indexed dataframes. 
-    Those dataframes are useful for simulating update operations.
-
-    The dataframes that are meant to be written need to be in different symbols for each of the processes
-    that asv creates. Thus the aim of the utilities is to help generation of the needed cache 
-    of dataframes in `setup_cache` as the actual writing should happen in asv `setup` method, where each method
-    ideally should use process id as identification for symbol, library etc.
-
-    Typically, you would generate four or more dataframes. The first dataframe initiates writing to a symbol,
-    while subsequent dataframes allow you to append by popping out from a list.
-    """
-
-    # Default frequency
-    FREQ = 's'
-    # Default initial timestamp
-    INITIAL_TIMESTAMP = pd.Timestamp("1999-10-10")
-
-    def __init__(self, storage, prefix = None, library_options = None, uris_cache = None):
-        super().__init__(storage, prefix, library_options, uris_cache)
-        self._frequency = self.FREQ
-        # Internally we keep timestamp as timestamp number for quick calculations
-        self.__init_time_number = TimestampNumber.from_timestamp(self.INITIAL_TIMESTAMP, self._frequency)
-        self.__number_cols = 20
-        self.__min_wide_col_threshold_columns = 500
-
-    def set_initial_timestamp(self, initial_timestamp):
-        self.__init_time_number = TimestampNumber.from_timestamp(initial_timestamp, self._frequency)
-        return self
-
-    def set_frequency(self, freq):
-        self._frequency = freq
-        return self
-
-    def set_default_columns(self, number_cols):
-        self.__number_cols = number_cols
-        return self
-
-    def get_default_columns(self) -> int:
-        return self.__number_cols
-
-    def generate_dataframe_no_index(self, number_rows: int):
-        """
-        In case needed no index dataframe is also possible to generate
-        """
-        number_cols = self.__number_cols
-        df: pd.DataFrame = DFGenerator.generate_random_dataframe(rows=number_rows, cols=number_cols, indexed=False)
-        return df
-    
-    def get_initial_time_number(self):
-        return copy.deepcopy(self.__init_time_number)
-
-    def set_index(self, df: pd.DataFrame , start_timestamp: Union[pd.Timestamp, TimestampNumber]):
-        if isinstance(start_timestamp, TimestampNumber):
-            start = start_timestamp.to_timestamp()
-        else:
-            start = start_timestamp
-        df.index = pd.date_range(start=start, 
-                                 periods=df.shape[0], freq=self.FREQ, name='timestamp')
-        return df
-
-    def generate_dataframe(self, number_rows: int, start_timestamp: Union[pd.Timestamp, TimestampNumber]):
-        
-        if self.__number_cols < self.__min_wide_col_threshold_columns:
-            self.logger().info("Dataframe generation started.")
-            st = time.time()
-            df = self.generate_dataframe_no_index(number_rows)
-            df = self.set_index(df, start_timestamp=start_timestamp)
-            self.logger().info(
-                f"Dataframe rows {number_rows} cols {self.__number_cols} generated for {time.time() - st} sec")
-        else:
-            df = self.generate_wide_dataframe(number_rows, start_timestamp)
-
-        return df
-    
-    def generate_wide_dataframe(self, number_rows: int, start_timestamp: Union[pd.Timestamp, TimestampNumber]):
-        """
-        Efficient for wide dataframes
-        """
-        st = time.time()
-        max_string_cols = 500
-        calculated_max = self.get_default_columns() // 15
-        if calculated_max > max_string_cols:
-            calculated_max =  max_string_cols
-
-        if isinstance(start_timestamp, TimestampNumber):
-            start = start_timestamp.to_timestamp()
-        else:
-            start = start_timestamp
-
-        df = DFGenerator.generate_wide_dataframe(num_rows=number_rows, num_cols=self.__number_cols,
-                                                 num_string_cols= calculated_max,
-                                                 start_time=start)
-        self.logger().info(f"Wide frame generated for {time.time() - st}")
-        return df
-
-    def generate_chained_writes(self, number_rows: int, 
-                               number_data_frames: int, 
-                               start_timestamp: pd.Timestamp = None,
-                               freq: str = 's') -> List[pd.DataFrame]:
-        """
-        Generates specified number of data frames each having specified number of rows.
-        The dataframes are in chronological order one after the other.
-        Date range starts with the specified initial timestamp setup and frequency
-        is the one setup
-
-        NOTE: will detect if the number of columns is 250 and will use another algorithm
-        for generating wide dataframe
-        """
-        cache = []
-        if start_timestamp:
-            timestamp_number = TimestampNumber.from_timestamp(start_timestamp, freq)
-        else: 
-            timestamp_number = self.get_initial_time_number()
-
-        for i in range(number_data_frames):
-            if self.get_default_columns() < self.__min_wide_col_threshold_columns:
-                df = self.generate_dataframe(number_rows, timestamp_number)
-            else: 
-                df = self.generate_wide_dataframe(number_rows, timestamp_number)
-            cache.append(df)
-            timestamp_number.inc(df.shape[0])
-        
-        return cache
-    
-    def get_first_and_last_timestamp(self, sequence_df_list: List[pd.DataFrame]) -> List[pd.Timestamp]:
-        """
-        Returns first and last timestamp of the list of indexed dataframes
-        """
-        assert len(sequence_df_list) > 0
-        start = sequence_df_list[0].index[0]
-        last = sequence_df_list[-1].index[-1]
-        return (start, last)
-    
-    def get_next_timestamp_number(self, sequence_df_list: List[pd.DataFrame], freq: str) -> TimestampNumber:
-        """
-        Returns next timestamp after the last timestamp in passed sequence of 
-        indexed dataframes.
-        """
-        last = self.get_first_and_last_timestamp(sequence_df_list)[1]
-        next = TimestampNumber.from_timestamp(last, freq) + 1
-        return next
-    
-
-class SetupMultipleLibraries(EnvConfigurationBase):
-    """
-        Sets up multiple libraries, each containing specified number of symbols
-        and each symbols having specified row numbers (col numbers can be defined also)
-        Accepted ASV params:
-          - list number of symbols per library
-          - number of rows in each list
-
-        The number of libraries to be created will depend on length of first list
-    """
-
-    def __init__(self, storage, prefix = None, library_options = None, uris_cache = None):
-        super().__init__(storage, prefix, library_options, uris_cache)
-        self.default_number_cols = 10
-        self.start_timestamp = pd.Timestamp("2000-1-1")
-        self.index_freq = 's'
-        self.indexed = True
-
-    def set_params(self, params) -> 'SetupMultipleLibraries':
-        assert len(params) == 2, "Supported 2 parameters"
-        super().set_params(params)
-        return self
-
-    def set_number_columns(self, num_cols:int = 2) -> 'SetupMultipleLibraries':
-        """
-        Sets the index of the column number parameter, set to None if all symbols
-        should have same number of columns, which is `default_number_cols`
-        """
-        self.default_number_cols = num_cols
-        return self
-    
-    def set_no_index(self) -> 'SetupMultipleLibraries':
-        self.indexed = False
-        return self
-    
-    def set_timestamp(self, timestamp: pd.Timestamp) -> 'SetupMultipleLibraries':
-        self.start_timestamp  = timestamp
-        return self
-
-    def generate_dataframe(self, rows: int, cols: int) -> pd.DataFrame:
-        """
-        Dataframe generator that will be used for wide dataframe generation
-        """
-        if cols is None:
-            cols = 1
-        st = time.time()
-        self.logger().info("Dataframe generation started.")
-        df: pd.DataFrame = DFGenerator.generate_random_dataframe(rows=rows, cols=cols, indexed=False)
-        if self.indexed:
-            df.index = pd.date_range(start=self.start_timestamp, 
-                                     periods=df.shape[0], freq=self.index_freq, name='primary_index')
-        self.logger().info(f"Dataframe rows {rows} cols {cols} generated for {time.time() - st} sec")
-        return df
-    
-    def get_symbol_name(self, symbol_number, rows, cols) -> str:
-        '''
-        Maps rows and cols to proper symbol name
-        This method should be used in ASV tests obtain symbol name based on parameters
-        '''
-        sym_suffix = ""
-        if cols is None:
-            sym_suffix = f"_idx{symbol_number}_rows{rows}"
-        else:
-            sym_suffix = f"_idx{symbol_number}_rows{rows}_cols{cols}"
-        return self.get_symbol_name_template(sym_suffix)
-    
-    def setup_symbol(self, lib: Library, symbols_number: int):
-        start = time.time()           
-        for counter in range(symbols_number):
-            for row in self._params[1]: #nd parameter is always number of rows
-                st = time.time()
-                cols = self.default_number_cols
-                df = self.generate_dataframe(row, cols)
-                symbol = self.get_symbol_name(counter, row, cols)
-                lib.write(symbol, df)
-                self.logger().debug(f"Dataframe stored at {lib} : {symbol} for {time.time() - st} sec")
-
-        self.logger().info(f"Stored {symbols_number} symbols for {time.time() - start} sec")
-
-
-    def setup_all(self) -> 'SetupMultipleLibraries':
-        """
-        Sets up in default library a number of symbols which is combination of parameters
-        for number of rows and columns
-        For each of symbols it will generate dataframe based on `generate_dataframe` function
-        """
-        start = time.time()
-        for num_symbols in self._params[0]:
-            lib = self.get_library(num_symbols)
-            self.logger().info(f"Library: {lib}")
-            self.setup_symbol(lib,num_symbols)
-        self.logger().info(f"TOTAL TIME (setup of one library with specified rows and column symbols): {time.time() - start} sec")
-        return self
-    
-    def check_ok(self):
-        """
-        Checks if library contains all needed data to run tests.
-        if OK, setting things up can be skipped
-        """
- 
-        for num_symbols in self._params[0]: # The number of rows param
-            if not self.has_library(num_symbols):
-                return False
-            lib = self.get_library(num_symbols)
-            self.logger().debug(f"Library {lib}")
-            for num_symbol in range(num_symbols):
-                symbols = set(lib.list_symbols())
-                self.logger().debug(f"Symbols {symbols}")
-                for rows in self._params[1]: # The number of rows param
-                    cols = self.default_number_cols
-                    symbol = self.get_symbol_name(num_symbol, rows, cols)
-                    self.logger().debug(f"Check symbol {symbol}")
-                    if not symbol in symbols:
-                        self.logger().info(f"CHECK FAILED")
-                        return False
-        self.logger().info(f"CHECK ENVIRONMENT PASSED ALL IS THERE")
-        return True
-
-
-class TestsClassForSetupEnvironmentFramework:
-    """
-    This set of tests allows to test the setup / check / teardown logic of setup environment 
-    building block classes 
-    """
-
-    PREFIX_FOR_TEST = "SOME_PREFIX"
-
-    @classmethod
-    def get_general_modifyanle_client(cls):
-        setup = (SetupSingleLibrary(Storage.AMAZON, 
-                                                prefix=TestsClassForSetupEnvironmentFramework.PREFIX_FOR_TEST))
-        assert TestsClassForSetupEnvironmentFramework.PREFIX_FOR_TEST in setup.get_arctic_client_modifiable().get_uri()
-        return setup
-
-
-    @classmethod
-    def get_general_setup_test_mode(cls):
-        setup = (SetupSingleLibrary(Storage.LMDB)
-                 .set_test_mode())
-        assert setup.is_test_mode()
-        return setup
-
-    @classmethod
-    def delete_test_store(cls, setup: EnvConfigurationBase):
-        ac = setup.get_arctic_client_persistent()
-        assert setup.is_test_mode()
-        libs = ac.list_libraries()
-        for lib in libs:
-            ac.delete_library(lib)
-        libs = ac.list_libraries()
-        assert len(libs) < 1, f"All libs should be deleted: {libs}"
-
-    @classmethod
-    def test_setup_with_rows_and_cols(cls):
-        setup = cls.get_general_setup_test_mode()
-        setup.set_params([[20,30],[40,50]])
-        df = setup.generate_dataframe(10,20)
-        assert df.shape[0] == 10
-        assert df.shape[1] == 20
-        logger.info(df)
-        cls.delete_test_store(setup)
-        if not setup.check_ok():
-            setup.setup_all()
-        assert setup.check_ok()
-        cls.delete_test_store(setup)
-        setup.setup_environment()
-        assert setup.check_ok()
-        cls.delete_test_store(setup)
-
-    @classmethod
-    def test_modifiable_client_workflow(cls):
-        """
-        Test for general functions for obtaining and 
-        cleaning modifiable libraries
-        """
-        setup = cls.get_general_modifyanle_client()
-        symbol = "test_symbol"
-        lib = setup.get_modifiable_library()
-        lib.write(symbol, setup.generate_dataframe(10,10))
-        ac = setup.get_arctic_client_modifiable()
-        assert ac.get_library(setup.get_library_name(LibraryType.MODIFIABLE))
-        setup.delete_modifiable_library()
-        assert len(ac.list_libraries()) == 0
-        assert len(setup.get_arctic_client_modifiable().list_libraries()) == 0
-
-    @classmethod
-    def test_setup_versions_and_snapshots(cls):
-        setup = (SetupLibrariesWithVersionAndSnapshots(storage=Storage.LMDB, prefix="LIST_SYMBOLS")
-        .set_with_metadata_for_each_version()
-        .set_with_snapshot_for_each_version()
-        .set_params([10, 20])
-        .set_test_mode())
-        assert setup.is_test_mode()
-
-    	#Delete-setup-all-check
-        cls.delete_test_store(setup)
-        setup.setup_environment()
-        setup.logger().info(setup.get_arctic_client_persistent().list_libraries())
-        assert setup.check_ok()
-
-        #Changing parameters should trigger not ok for setup
-        setup.set_with_metadata_for_each_version()
-        setup.set_with_snapshot_for_each_version()
-        setup.set_params([2, 3])
-        assert not setup.check_ok()
-
-    @classmethod
-    def test_setup_multiple_libraries_with_symbols(cls):
-        """
-        Tests for `SetupMultipleLibraries` correctness
-        (not used in ASV)
-        """
-        params = [[2, 3], [4, 5]]
-        cols = 12
-        setup = (SetupMultipleLibraries(storage=Storage.LMDB, 
-                                                      prefix="BASIC_BATCH")
-                   .set_params(params) # For test purposes
-                   .set_number_columns(cols)
-                   ).set_test_mode()
-        assert setup.is_test_mode()            
-        
-        cls.delete_test_store(setup)
-
-        assert not setup.check_ok()
-        setup.setup_environment()
-        assert setup.check_ok()
-
-        ac = setup.get_arctic_client_persistent()
-        libs = ac.list_libraries()
-        setup.logger().info(f"LIBS {libs}")
-        assert len(libs) == len(params[0]) # the symbol list
-
-        for num_syms in params[0]: # the symbols list
-            ln = setup.get_library_name(LibraryType.PERSISTENT, num_syms)
-            lib = ac.get_library(ln)
-            symbols_list = lib.list_symbols()
-            setup.logger().info(symbols_list)
-            for rows in params[1]: # the symbols list
-                setup.get_symbol_name(num_syms, rows, cols)
-                assert setup.has_library(num_syms) 
-                lib = setup.get_library(num_syms)
-                symbol = setup.get_symbol_name(num_syms - 1, rows, cols)
-                assert lib.read(symbol).data is not None
